Make JIRA ticket match case insensitive
Previously, we only fetched JIRA ticket if it was on the format "JIRA:
<ticket>", now the "JIRA:" part is case insensitive.

Also made it so that if no ticket was found, that part of the title gets
removed entirely instead of being "None". This lead to a minor refactor
of this part of the code, which should make it cleaner to add/remove
parts in the future.