Add a subject line and extend notifications
The name of the author and the title of the program is now added to the
subject line of of the notification and will typically show up in bold
text above the body, which consists of the new comment. The expiration
time of the notification is also extended to 15 seconds.