Introduce --no-notifications flag
When this flag is used, reviewcheck will not attempt to invoke
`notify-send(1)`. This is required to be able to run reviewcheck in an
environment that a) does not have notify-send installed, or b) is not
the host machine (e.g. VM, container).

Whether or not this solution is the best (or if sending notifications is
a sane default) is still uncertain to me.