Show MRs where user has reacted but not upvoted
If a user has left a reaction on a merge request but not upvoted it,
that merge request will now be shown in the feed. It is considered
active for the user. It can be hidden from the feed by either removing
the non-thumbs up reaction on the merge request, or upvoting it with a
thumbs up. If the user has active threads on the merge request, it is
shown anyway. This new functionality just shows info boxes for merge
requests that would otherwise be hidden, with no threads underneath.