Use parameter flag instead of inline flag in regex
Using (?i) inside a regex is deprecated. It is still possible to use it
at the start of the regex. I consider it clearer to have the flags as
parameters instead of inside the regex string. So I set the
re.IGNORECASE parameter flag instead.

Also added a test for the function where the regex is used.